pyproject.toml
[project]
name = "WEG-Solutions"
version = "0.1.0"
requires-python = "==3.13"
dependencies = [
    "gunicorn",
    "gevent",
    "django",
    "python-environ",
    "django-compressor",
    "django-allauth",
    "whitenoise",
    "celery[redis]",
    "django-celery-beat",
    "psycopg[binary]",
    "sentry-sdk",
]

[dependency-groups]
dev = [
    "ruff",
    "pre-commit",
    "uv",
    "pytest-django",
    "factory-boy",
    "mock",
    "watchfiles",
    "freezegun",
    "djlint",
    "pytest-cov",
    "ipython",
]


[tool.pytest.ini_options]
python_files = "tests.py test_*.py *_tests.py"
addopts = "--ds=weg_solutions.settings_test --reuse-db"

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "Q", "PT", "C4", "B", "A"]
ignore = ["B904"]
